pyproject.toml
[project]
name = "rtdetr-inference"
version = "0.1.0"
description = "RT-DETR ONNX inference toolkit"
requires-python = ">=3.8"
dependencies = [
    "numpy>=1.21.0",
    "opencv-python>=4.5.0",
    "onnxruntime>=1.15.0",
    "onnxruntime-gpu>=1.20.2",
    "onnx>=1.17.0",
    "onnxruntime-tools>=1.7.0",
    "torch>=2.5.1",
    "torchvision>=0.20.1",
    "transformers>=4.46.3",
    "pillow>=10.4.0",
    "requests>=2.32.4",
    "scipy>=1.10.1",
    "optimum[onnxruntime]>=1.23.3",
]

[project.optional-dependencies]
dev = [
    "onnx>=1.12.0",
    "onnxsim>=0.4.0",
    "transformers>=4.20.0",
]
